+++ src/SprykerEco/Zed/Unzer/Business/Payment/Processor/Refund/MarketplaceRefundProcessor.php
@@ -198,7 +198,7 @@ class MarketplaceRefundProcessor implements UnzerRefundProcessorInterface
     {
         return (new UnzerRefundItemTransfer())
             ->setParticipantId($itemTransfer->getUnzerParticipantId())
-            ->setAmountGross((int)$itemTransfer->getRefundableAmount() / UnzerConstants::INT_TO_FLOAT_DIVIDER)
+            ->setAmountGross((float)$itemTransfer->getRefundableAmount() / UnzerConstants::INT_TO_FLOAT_DIVIDER)
             ->setBasketItemReferenceId($itemTransfer->getSku())
             ->setQuantity(UnzerConstants::PARTIAL_REFUND_QUANTITY);
     }
